adding user

+++ user/user.h
@@ -26,7 +26,7 @@ int sleep(int);
 int uptime(void);
 int getprocs(struct pstat*);
 //Adding declaration
-uint64 freepmem(void);
+int freepmem(void);
 
 // ulib.c
 int stat(const char*, struct stat*);
